HBASE-23165 [hbtop] Some modifications from HBASE-22988 (#987)
