+++ hbase-hbtop/src/test/java/org/apache/hadoop/hbase/hbtop/terminal/impl/TestCursor.java
@@ -15,20 +15,53 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.hbase.hbtop.terminal;
+package org.apache.hadoop.hbase.hbtop.terminal.impl;
 
 import java.util.concurrent.TimeUnit;
-import org.apache.hadoop.hbase.hbtop.terminal.impl.TerminalImpl;
+
+import org.apache.hadoop.hbase.hbtop.terminal.KeyPress;
+import org.apache.hadoop.hbase.hbtop.terminal.Terminal;
 
 
-public final class KeyPressTest {
+public final class TestCursor {
 
-  private KeyPressTest() {
+  private TestCursor() {
   }
 
   public static void main(String[] args) throws Exception {
     try (Terminal terminal = new TerminalImpl()) {
-      terminal.hideCursor();
+      terminal.refresh();
+      terminal.setCursorPosition(0, 0);
+
+      terminal.getTerminalPrinter(0).print("aaa").endOfLine();
+      terminal.refresh();
+      TimeUnit.SECONDS.sleep(1);
+
+      terminal.getTerminalPrinter(0).print("bbb").endOfLine();
+      terminal.refresh();
+      TimeUnit.SECONDS.sleep(1);
+
+      terminal.setCursorPosition(1, 0);
+      terminal.refresh();
+      TimeUnit.SECONDS.sleep(1);
+
+      terminal.setCursorPosition(2, 0);
+      terminal.refresh();
+      TimeUnit.SECONDS.sleep(1);
+
+      terminal.setCursorPosition(3, 0);
+      terminal.refresh();
+      TimeUnit.SECONDS.sleep(1);
+
+      terminal.setCursorPosition(0, 1);
+      terminal.refresh();
+      TimeUnit.SECONDS.sleep(1);
+
+      terminal.getTerminalPrinter(1).print("ccc").endOfLine();
+      terminal.refresh();
+      TimeUnit.SECONDS.sleep(1);
+
+      terminal.getTerminalPrinter(3).print("Press any key to finish").endOfLine();
       terminal.refresh();
 
       while (true) {
@@ -37,13 +70,7 @@ public final class KeyPressTest {
           TimeUnit.MILLISECONDS.sleep(100);
           continue;
         }
-
-        terminal.getTerminalPrinter(0).print(keyPress.toString()).endOfLine();
-        terminal.refresh();
-
-        if (keyPress.getType() == KeyPress.Type.F12) {
-          break;
-        }
+        break;
       }
     }
   }
